@@ -19,8 +19,8 @@ package com.netflix.spinnaker.kork.retrofit.exceptions;
 import com.netflix.spinnaker.kork.web.exceptions.BaseExceptionHandlers;
 import com.netflix.spinnaker.kork.web.exceptions.ExceptionMessageDecorator;
 import java.io.IOException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.core.Ordered;
