@@ -15,20 +15,14 @@
  */
 package com.netflix.spinnaker.kork.jedis;
 
-import java.util.List;
-import java.util.function.Consumer;
-import java.util.function.Function;
 import redis.clients.jedis.Jedis;
 import redis.clients.jedis.Pipeline;
-import redis.clients.jedis.ScanParams;
-import redis.clients.jedis.ScanResult;
 import redis.clients.jedis.Transaction;
-import redis.clients.jedis.commands.BinaryJedisCommands;
 import redis.clients.jedis.commands.JedisCommands;
-import redis.clients.jedis.commands.MultiKeyCommands;
-import redis.clients.jedis.commands.RedisPipeline;
-import redis.clients.jedis.commands.ScriptingCommands;
 import redis.clients.jedis.util.Pool;
+
+import java.util.function.Consumer;
+import java.util.function.Function;
 
 public class JedisClientDelegate implements RedisClientDelegate {
 
@@ -63,14 +57,14 @@ public class JedisClientDelegate implements RedisClientDelegate {
     }
   }
 
-  @Override
+  /*@Override
   public <R> R withMultiClient(Function<MultiKeyCommands, R> f) {
     try (Jedis jedis = jedisPool.getResource()) {
       return f.apply(jedis);
     }
-  }
+  }*/
 
-  @Override
+  /*@Override
   public void withMultiClient(Consumer<MultiKeyCommands> f) {
     try (Jedis jedis = jedisPool.getResource()) {
       f.accept(jedis);
@@ -113,7 +107,7 @@ public class JedisClientDelegate implements RedisClientDelegate {
     }
 
     ((Pipeline) p).sync();
-  }
+  }*/
 
   @Override
   public boolean supportsMultiKeyPipelines() {
@@ -159,6 +153,11 @@ public class JedisClientDelegate implements RedisClientDelegate {
   }
 
   @Override
+  public void withKeyScan(String pattern, int count, Consumer<RedisScanResult> f) {
+
+  }
+
+  /*@Override
   public void withScriptingClient(Consumer<ScriptingCommands> f) {
     try (Jedis jedis = jedisPool.getResource()) {
       f.accept(jedis);
@@ -187,5 +186,5 @@ public class JedisClientDelegate implements RedisClientDelegate {
         cursor = result.getCursor();
       } while (!"0".equals(cursor));
     }
-  }
+  }*/
 }
