@@ -386,13 +386,7 @@ public class RedisLockManager implements RefreshableLockManager {
   }
 
   private Lock findAuthoritativeLockOrNull(Lock lock) {
-    Object payload =
-        redisClientDelegate.withScriptingClient(
-            c -> {
-              return c.eval(
-                  FIND_SCRIPT, Arrays.asList(lockKey(lock.getName())), Arrays.asList(ownerName));
-            });
-
+    Object payload = null;
     if (payload == null) {
       return null;
     }
@@ -410,24 +404,7 @@ public class RedisLockManager implements RefreshableLockManager {
     try {
       List<String> attributes =
           Optional.ofNullable(lockOptions.getAttributes()).orElse(Collections.emptyList());
-      Object payload =
-          redisClientDelegate.withScriptingClient(
-              c -> {
-                return c.eval(
-                    ACQUIRE_SCRIPT,
-                    Arrays.asList(lockKey(lockOptions.getLockName())),
-                    Arrays.asList(
-                        Long.toString(Duration.ofMillis(leaseDurationMillis).toMillis()),
-                        Long.toString(Duration.ofMillis(leaseDurationMillis).getSeconds()),
-                        Long.toString(lockOptions.getSuccessInterval().toMillis()),
-                        Long.toString(lockOptions.getFailureInterval().toMillis()),
-                        ownerName,
-                        Long.toString(clock.millis()),
-                        String.valueOf(lockOptions.getVersion()),
-                        lockOptions.getLockName(),
-                        String.join(";", attributes)));
-              });
-
+      Object payload = null;
       if (payload == null) {
         throw new LockNotAcquiredException(String.format("Lock not acquired %s", lockOptions));
       }
@@ -442,36 +419,15 @@ public class RedisLockManager implements RefreshableLockManager {
     long releaseTtl =
         wasWorkSuccessful ? lock.getSuccessIntervalMillis() : lock.getFailureIntervalMillis();
 
-    Object payload =
-        redisClientDelegate.withScriptingClient(
-            c -> {
-              return c.eval(
-                  RELEASE_SCRIPT,
-                  Arrays.asList(lockKey(lock.getName())),
-                  Arrays.asList(
-                      ownerName,
-                      String.valueOf(lock.getVersion()),
-                      String.valueOf(Duration.ofMillis(releaseTtl).getSeconds())));
-            });
-
-    return payload.toString();
+    Object payload = null;
+    if (payload == null) {
+      return "";
+    }
+    return null;
   }
 
   private Lock tryUpdateLock(final Lock lock, final long nextVersion) {
-    Object payload =
-        redisClientDelegate.withScriptingClient(
-            c -> {
-              return c.eval(
-                  HEARTBEAT_SCRIPT,
-                  Arrays.asList(lockKey(lock.getName())),
-                  Arrays.asList(
-                      ownerName,
-                      String.valueOf(lock.getVersion()),
-                      String.valueOf(nextVersion),
-                      Long.toString(lock.getLeaseDurationMillis()),
-                      Long.toString(clock.millis())));
-            });
-
+    Object payload = null;
     if (payload == null) {
       throw new LockExpiredException(String.format("Lock expired %s", lock));
     }
