@@ -29,7 +29,9 @@ import com.netflix.spinnaker.kork.plugins.remote.extension.transport.RemoteExten
 import com.netflix.spinnaker.security.AuthenticatedRequest
 import okhttp3.Headers
 import okhttp3.HttpUrl
+import okhttp3.HttpUrl.Companion.toHttpUrlOrNull
 import okhttp3.MediaType
+import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.RequestBody
@@ -56,7 +58,7 @@ class OkHttpRemoteExtensionTransport(
         .headers(buildHeaders(httpConfig.headers.invokeHeaders))
         .post(
           RequestBody.create(
-            MediaType.parse("application/json"),
+            "application/json".toMediaTypeOrNull(),
             objectMapper.writeValueAsString(remoteExtensionPayload)
           )
         )
@@ -64,7 +66,7 @@ class OkHttpRemoteExtensionTransport(
 
       val response = client.newCall(request).execute()
       if (!response.isSuccessful) {
-        val reason = response.body()?.string() ?: "Unknown reason: ${response.code()}"
+        val reason = response.body?.string() ?: "Unknown reason: ${response.code}"
         throw OkHttpRemoteExtensionTransportException(reason)
       }
     }.call()
@@ -77,7 +79,7 @@ class OkHttpRemoteExtensionTransport(
         .headers(buildHeaders(httpConfig.headers.writeHeaders))
         .post(
           RequestBody.create(
-            MediaType.parse("application/json"),
+            "application/json".toMediaTypeOrNull(),
             objectMapper.writeValueAsString(remoteExtensionPayload)
           )
         )
@@ -85,11 +87,11 @@ class OkHttpRemoteExtensionTransport(
 
       val response = client.newCall(request).execute()
       if (!response.isSuccessful) {
-        val reason = response.body()?.string() ?: "Unknown reason: ${response.code()}"
+        val reason = response.body?.string() ?: "Unknown reason: ${response.code}"
         throw OkHttpRemoteExtensionTransportException(reason)
       }
 
-      objectMapper.readValue(response.body()?.string(), RemoteExtensionResponse::class.java)
+      objectMapper.readValue(response.body?.string(), RemoteExtensionResponse::class.java)
     }.call()
   }
 
@@ -103,16 +105,16 @@ class OkHttpRemoteExtensionTransport(
 
       val response = client.newCall(request).execute()
       if (!response.isSuccessful) {
-        val reason = response.body()?.string() ?: "Unknown reason: ${response.code()}"
+        val reason = response.body?.string() ?: "Unknown reason: ${response.code}"
         throw OkHttpRemoteExtensionTransportException(reason)
       }
 
-      objectMapper.readValue(response.body()?.string(), RemoteExtensionResponse::class.java)
+      objectMapper.readValue(response.body?.string(), RemoteExtensionResponse::class.java)
     }.call()
   }
 
   private fun buildUrl(additionalParams: Map<String, String>): HttpUrl {
-    val httpUrlBuilder = HttpUrl.parse(httpConfig.url)?.newBuilder()
+    val httpUrlBuilder = httpConfig.url.toHttpUrlOrNull()?.newBuilder()
       ?: throw IntegrationException("Unable to parse url '${httpConfig.url}'")
 
     (httpConfig.queryParams + additionalParams).forEach {
